+++ blueprints/molecule-capsule/m-chemx-badge/index.ts
@@ -1,3 +1,3 @@
 export * from './types.d';
-export { MChemxBadge } from './m-chemx-badge';
+export { MChemxBadge, resolveGradeClass } from './m-chemx-badge';
 export { default } from './m-chemx-badge';